commenting out hard to read areas
