+++ src/config/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export const baseUrl = 'https://quickdecks.herokuapp.com/api';
+export const baseUrl = 'https://quickdecks-staging.herokuapp.com/api';
