+++ src/modules/user/userActions.js
@@ -1,4 +1,9 @@
+// Import
+
+// Libraries
 import axios from 'axios';
+
+// Types
 import {
   LOADING,
   LOGIN,
@@ -7,7 +12,11 @@ import {
   RESET_PASSWORD,
   FORGOT_PASSWORD,
 } from './userTypes';
+
+// Configs
 import { baseUrl } from '../../config/index';
+
+// Utils
 import { axiosWithAuth } from '../../utils/auth';
 
 export const userLogin = (email, password, history) => dispatch => {
@@ -21,7 +30,7 @@ export const userLogin = (email, password, history) => dispatch => {
       dispatch({
         type: LOGIN,
       });
-      localStorage.setItem('token', `${data.token}`);
+      localStorage.setItem('token', `${data.data.token}`);
       history.push('/dashboard');
     });
 };
@@ -30,15 +39,9 @@ export const userSignUp = (userData, history) => dispatch => {
   dispatch({ type: LOADING });
   axios
     .post(`${baseUrl}/auth/register`, userData)
-    .then(() => {
-      dispatch(
-        userLogin(
-          userData.fullName,
-          userData.password,
-          userData.email,
-          history,
-        ),
-      );
+    .then(({ data }) => {
+      localStorage.setItem('token', `${data.data.token}`);
+      history.push('/dashboard');
     })
     .catch(errors => {
       dispatch({
