+++ src/pages/login/Login.js
@@ -1,16 +1,13 @@
+// Imports
+
+// Libraries
 import React from 'react';
 import { NavLink, Redirect } from 'react-router-dom';
-import { withFormik } from 'formik';
-import { connect } from 'react-redux';
 import { useHistory } from 'react-router';
-import styled from 'styled-components';
 
-import * as yup from 'yup';
-import * as c from '../../styles/variables/colours';
-import { userLogin } from '../../modules/user/userActions';
-
-import { H1, Text, H3, H5 } from '../../styles/typography';
-import { Button, BackArrowButton } from '../../styles/buttons';
+// Styles
+import { H1, H3, H5 } from '../../styles/typography';
+import { BackArrowButton } from '../../styles/buttons';
 import {
   ParentBackground,
   SkewDiv,
@@ -20,27 +17,16 @@ import {
   FlexRowBackground,
   DesktopImage,
 } from '../../styles/background';
-import { Forms, Input, Label } from '../../styles/forms';
 
+// Assets
 import img from '../../assets/images/undraw_online_test_gba7 (1).svg';
 import BackArrow from '../../assets/icons/Arrow 1.svg';
 
-const LoginForm = props => {
-  const ForgotText = styled(H3)`
-    align-self: flex-end;
-    line-height: 0;
-    margin-bottom: 1em;
-  `;
+// Components
+import LoginForm from '../../components/loginForm/LoginForm';
 
+export default function Login() {
   const history = useHistory();
-  const {
-    values,
-    handleChange,
-    handleBlur,
-    handleSubmit,
-    touched,
-    errors,
-  } = props;
   return (
     <FlexRowBackground>
       {localStorage.getItem('token') && <Redirect to="/dashboard" />}
@@ -60,85 +46,15 @@ const LoginForm = props => {
         <SkewDiv>
           <UnSkewDiv>
             <H1>Login</H1>
-            <Forms onSubmit={handleSubmit}>
-              <Label>
-                <H3>Email</H3>
-                {touched.email && errors.email && (
-                  <Text color={c.DANGER_COLOR}>{errors.email}</Text>
-                )}
-                <Input
-                  name="email"
-                  value={values.email}
-                  onChange={handleChange}
-                  onBlur={handleBlur}
-                  placeholder="Email"
-                  border={
-                    touched.email &&
-                    errors.email &&
-                    `2px solid ${c.DANGER_COLOR}`
-                  }
-                />
-              </Label>
-              <Label>
-                <H3>Password</H3>
-                {touched.password && errors.password && (
-                  <Text color={c.DANGER_COLOR}>{errors.password}</Text>
-                )}
-                <Input
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  value={values.password}
-                  onBlur={handleBlur}
-                  onChange={handleChange}
-                  border={errors.password && `2px solid ${c.DANGER_COLOR}`}
-                />
-              </Label>
-              <ForgotText REGULAR>
-                <NavLink to="/forgot">Forgot Password?</NavLink>
-              </ForgotText>
-              <Button type="submit">
-                <H3 WHITE>Login</H3>
-              </Button>
-              <br />
-              <H3 REGULAR>
-                Do not have an account? <NavLink to="/signup">Sign Up</NavLink>
-              </H3>
-            </Forms>
+            <LoginForm />
+            <br />
+            <H3 REGULAR>
+              Do not have an account? <NavLink to="/signup">Sign Up</NavLink>
+            </H3>
           </UnSkewDiv>
         </SkewDiv>
         <BottomTriangle />
       </ParentBackground>
     </FlexRowBackground>
   );
-};
-
-const validationSchema = yup.object().shape({
-  email: yup
-    .string()
-    .email('Email is not valid')
-    .required('Please provide an email'),
-  password: yup
-    .string()
-    .required('Please provide a password')
-    .min(8, 'Password too short'),
-});
-
-const Login = withFormik({
-  mapPropsToValues: () => ({
-    email: '',
-    password: '',
-  }),
-  handleSubmit: (values, { props, setSubmitting }) => {
-    props.userLogin(values.email, values.password, props.history);
-    setSubmitting(false);
-  },
-  validationSchema,
-})(LoginForm);
-
-const mapStateToProps = state => {
-  return {
-    user: state.user,
-  };
-};
-export default connect(mapStateToProps, { userLogin })(Login);
+}
