+++ src/utils/PrivateRoute.js
@@ -1,3 +1,6 @@
+// Import
+
+// Libraries
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
